+++ src/hotspot/share/gc/shenandoah/shenandoahGenerationalControlThread.cpp
@@ -232,7 +232,7 @@ void ShenandoahGenerationalControlThread::run_gc_cycle(const ShenandoahGCRequest
 
   // GC is starting, bump the internal gc count and set GCIdMark
   update_gc_count();
-  GCIdMark gc_id_mark(get_gc_id());
+  GCIdMark gc_id_mark(static_cast<uint>(get_gc_id()));
 
   _heap->reset_bytes_allocated_since_gc_start();
